4.0.0/stability - Audiences panel issues - Adjust item mega-menu auto focus behaviours & accessibility
