+++ ui/app/src/components/Controls/FormEngine/styles.ts
@@ -14,19 +14,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { StyleRules } from '@material-ui/core/styles';
+import { createStyles, makeStyles } from '@material-ui/core/styles';
 
-export const controlBaseStyles: StyleRules<any, any> = {
-  formControl: {
-    width: '100%',
-    '& .MuiFormGroup-root': {
-      marginLeft: '10px'
-    },
-    '& .MuiInputBase-root': {
-      marginTop: '12px !important'
-    }
-  },
-  inputLabel: {
-    position: 'relative'
-  }
-};
+export const useStyles = makeStyles(() =>
+  createStyles({
+    formControl: {},
+    inputLabel: {}
+  })
+);
+
+export default useStyles;
